renamded index to export so it's obvious what the library "exports"
