+++ examples/create-fixture-person.test.ts
@@ -1,6 +1,6 @@
+import { Fixture } from '@/export';
 import { expect, test } from 'vitest';
 import { z } from 'zod';
-import { Fixture } from 'zod-fixture';
 
 test('generates a person', () => {
 	const PersonSchema = z.object({
